Resolving some warnings

+++ Firmware/src/Microchip/usb_device.c
@@ -1781,7 +1781,7 @@ static void USBCtrlTrfTxService(void)
     byteToSend = USB_EP0_BUFF_SIZE;
     if(inPipes[0].wCount.Val < (uint8_t)USB_EP0_BUFF_SIZE)
     {
-        byteToSend = inPipes[0].wCount.Val;
+        byteToSend = (uint8_t)inPipes[0].wCount.Val;
 
         //Keep track of whether or not we have sent a "short packet" yet.
         //This is useful so that later on, we can configure EP0 IN to STALL,
@@ -1868,7 +1868,7 @@ static void USBCtrlTrfRxService(void)
     //application firmware was expecting to receive.
     if(byteToRead > outPipes[0].wCount.Val)
     {
-        byteToRead = outPipes[0].wCount.Val;
+        byteToRead = (uint8_t)outPipes[0].wCount.Val;
     }
     //Reduce the number of remaining bytes by the number we just received.
     outPipes[0].wCount.Val -= byteToRead;
